Changed ASN sources. Added SIB1 packing/unpacking to RRC

+++ lte/rrc/asn/per_encoder.c
@@ -88,7 +88,7 @@ uper_encode_to_new_buffer(asn_TYPE_descriptor_t *td, asn_per_constraints_t *cons
 		}
 	default:
 		*buffer_r = key.buffer;
-		ASN_DEBUG("Complete encoded in %ld bits", (long)er.encoded);
+		ASN_DEBUG("Complete encoded in %d bits", er.encoded);
 		return ((er.encoded + 7) >> 3);
 	}
 }
